don hang

+++ app/Http/Controllers/DanhMucController.php
@@ -4,12 +4,23 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\DanhMucRequestCreate;
 use App\Models\DanhMuc;
+use App\Models\PhanQuyen;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class DanhMucController extends Controller
 {
     public function getOpenData()
     {
+        $id_chuc_nang = 21;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         $data = DanhMuc::where('tinh_trang',1)
                             ->get();;
         return response()->json([
@@ -19,6 +30,15 @@ class DanhMucController extends Controller
     }
     public function changeStatus(Request $request)
     {
+        $id_chuc_nang = 20;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         $danhMuc = DanhMuc::where('id', $request->id)->first();
         if (!$danhMuc) {
             return response()->json([
@@ -41,6 +61,15 @@ class DanhMucController extends Controller
     }
     public function store(DanhMucRequestCreate $request)
     {
+        $id_chuc_nang = 16;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         DanhMuc::create([
             'ten_danh_muc'   =>$request->ten_danh_muc,
             'tinh_trang'     =>$request->tinh_trang,
@@ -55,6 +84,15 @@ class DanhMucController extends Controller
 
     public function getData()
     {
+        $id_chuc_nang = 15;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         $data = DanhMuc::get();
 
         return response()->json([
@@ -64,6 +102,15 @@ class DanhMucController extends Controller
 
     public function update(Request $request)
     {
+        $id_chuc_nang = 17;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         DanhMuc::where('id', $request->id)->update([
             'ten_danh_muc'   =>$request->ten_danh_muc,
             'tinh_trang'     =>$request->tinh_trang,
@@ -78,6 +125,15 @@ class DanhMucController extends Controller
 
     public function destroy(Request $request)
     {
+        $id_chuc_nang = 18;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         DanhMuc::where('id', $request->id)->delete();
 
         return response()->json([
@@ -88,6 +144,15 @@ class DanhMucController extends Controller
 
     public function search(Request $request)
     {
+        $id_chuc_nang = 19;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         $noi_dung = '%' . $request->noi_dung . '%';
 
         $data = DanhMuc::where('ten_danh_muc', 'like', $noi_dung)
