+++ app/Http/Controllers/SanPhamShopController.php
@@ -5,112 +5,25 @@ namespace App\Http\Controllers;
 use App\Http\Requests\SanPhamRequestCreate;
 use App\Models\BienTheSanPham;
 use App\Models\HinhAnhSanPham;
+use App\Models\PhanQuyen;
 use App\Models\SanPham;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class SanPhamShopController extends Controller
 {
-    public function chiTietSanPham($id)
-    {
-        $data = DB::table('san_phams')
-            ->join('thuong_hieus', 'thuong_hieus.id', '=', 'san_phams.id_thuong_hieu')
-            ->join('danh_mucs', 'danh_mucs.id', '=', 'san_phams.id_danh_muc')
-            ->leftJoin('bien_the_san_phams', 'bien_the_san_phams.id_san_pham', '=', 'san_phams.id')
-            ->select(
-                'san_phams.id',
-                'thuong_hieus.ten_thuong_hieu',
-                'danh_mucs.id as id_danh_muc',
-                'danh_mucs.ten_danh_muc',
-                'san_phams.ten_san_pham',
-                'san_phams.slug_san_pham',
-                'san_phams.giam_gia',
-                'san_phams.mo_ta',
-                'san_phams.gia_goc',
-                DB::raw('COALESCE(SUM(bien_the_san_phams.so_luong_ton), 0) AS tong_sl_ton'),
-                DB::raw('san_phams.gia_goc - (san_phams.gia_goc * san_phams.giam_gia / 100) AS gia_ban')
-            )
-            ->where('san_phams.id', $id)
-            ->where('san_phams.tinh_trang', 1)
-            ->groupBy(
-                'san_phams.id',
-                'thuong_hieus.ten_thuong_hieu',
-                'danh_mucs.id',
-                'danh_mucs.ten_danh_muc',
-                'san_phams.ten_san_pham',
-                'san_phams.slug_san_pham',
-                'san_phams.mo_ta',
-                'san_phams.gia_goc',
-                'san_phams.giam_gia'
-            )
-            ->first();
-
-        if (!$data) {
-            return response()->json([
-                'status'  => false,
-                'message' => 'Sản phẩm không tồn tại hoặc không có sẵn.'
-            ]);
-        }
-
-        $images = DB::table('hinh_anh_san_phams')
-            ->where('id_san_pham', $id)
-            ->pluck('hinh_anh')
-            ->toArray();
-
-        $data->hinh_anh = $images;
-
-        $variants = DB::table('bien_the_san_phams')
-            ->where('id_san_pham', $id)
-            ->select('id', 'kich_thuoc', 'mau_sac')
-            ->get();
-
-        $data->variants = $variants;
-
-        return response()->json([
-            'status' => true,
-            'data' => $data
-        ]);
-    }
-
-    public function cungDanhMuc(Request $request)
-    {
-        $query = DB::table('san_phams')
-            ->join('danh_mucs', 'danh_mucs.id', '=', 'san_phams.id_danh_muc')
-            ->leftJoin('hinh_anh_san_phams', 'hinh_anh_san_phams.id_san_pham', '=', 'san_phams.id')
-            ->select(
-                'san_phams.id',
-                'san_phams.slug_san_pham',
-                'danh_mucs.ten_danh_muc',
-                'san_phams.ten_san_pham',
-                DB::raw('MIN(hinh_anh_san_phams.hinh_anh) AS hinh_anh'),
-                DB::raw('san_phams.gia_goc - (san_phams.gia_goc * san_phams.giam_gia / 100) AS gia_ban'),
-                'san_phams.gia_goc',
-                'san_phams.giam_gia'
-            )
-            ->where('danh_mucs.id', $request->id_danh_muc)
-            ->where('san_phams.tinh_trang', 1);
-
-        if ($request->has('id_san_pham')) {
-            $query->where('san_phams.id', '!=', $request->id_san_pham);
-        }
-
-        $data = $query->groupBy(
-            'san_phams.id',
-            'danh_mucs.ten_danh_muc',
-            'san_phams.ten_san_pham',
-            'san_phams.slug_san_pham',
-            'san_phams.gia_goc',
-            'san_phams.giam_gia'
-        )->get();
-
-        return response()->json([
-            'status' => true,
-            'data' => $data
-        ]);
-    }
-
     public function changeStatus(Request $request)
     {
+        $id_chuc_nang = 44;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         $sanPham = SanPham::where('id', $request->id)->first();
 
         if (!$sanPham) {
@@ -138,6 +51,15 @@ class SanPhamShopController extends Controller
 
     public function store(Request $request)
     {
+        $id_chuc_nang = 39;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         $sanPham = SanPham::create([
             'id_thuong_hieu' => $request->id_thuong_hieu,
             'id_danh_muc' => $request->id_danh_muc,
@@ -176,7 +98,7 @@ class SanPhamShopController extends Controller
                 'san_phams.giam_gia',
                 'san_phams.mo_ta',
                 'san_phams.gia_goc',
-                DB::raw('COALESCE(SUM(bien_the_san_phams.so_luong_ton), 0) AS TONG_SL_TON'),
+                DB::raw('COALESCE(SUM(bien_the_san_phams.so_luong_ton), 0) AS tong_sl_ton'),
                 DB::raw('(CASE WHEN COALESCE(SUM(bien_the_san_phams.so_luong_ton), 0) = 0 THEN 0 ELSE san_phams.tinh_trang END) AS tinh_trang'),
                 DB::raw('san_phams.gia_goc - (san_phams.gia_goc * san_phams.giam_gia / 100) AS gia_ban')
             )
@@ -193,7 +115,19 @@ class SanPhamShopController extends Controller
                 'san_phams.gia_goc',
                 'san_phams.giam_gia'
             )
+            ->inRandomOrder()
             ->get();
+
+        // Thêm logic lấy kích thước và màu sắc distinct
+        foreach ($data as $item) {
+            $variants = DB::table('bien_the_san_phams')
+                ->where('id_san_pham', $item->id)
+                ->select('kich_thuoc', 'mau_sac')
+                ->get();
+
+            $item->kich_thuoc_distinct = $variants->pluck('kich_thuoc')->unique()->values()->all();
+            $item->mau_sac_distinct = $variants->pluck('mau_sac')->unique()->values()->all();
+        }
 
         return response()->json([
             'data' => $data
@@ -202,6 +136,15 @@ class SanPhamShopController extends Controller
 
     public function getData(Request $request)
     {
+        $id_chuc_nang = 40;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         $data = DB::table('san_phams')
             ->join('thuong_hieus', 'thuong_hieus.id', '=', 'san_phams.id_thuong_hieu')
             ->join('danh_mucs', 'danh_mucs.id', '=', 'san_phams.id_danh_muc')
@@ -248,6 +191,15 @@ class SanPhamShopController extends Controller
 
     public function update(Request $request)
     {
+        $id_chuc_nang = 41;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         SanPham::where('id', $request->id)->update([
             'id_thuong_hieu' => $request->id_thuong_hieu,
             'id_danh_muc'    => $request->id_danh_muc,
@@ -267,6 +219,15 @@ class SanPhamShopController extends Controller
 
     public function destroy(Request $request)
     {
+        $id_chuc_nang = 42;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         SanPham::where('id', $request->id)->delete();
 
         return response()->json([
@@ -277,6 +238,15 @@ class SanPhamShopController extends Controller
 
     public function search(Request $request)
     {
+        $id_chuc_nang = 43;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         $noi_dung = '%' . $request->noi_dung . '%';
 
         $data = DB::table('san_phams')
@@ -295,7 +265,15 @@ class SanPhamShopController extends Controller
 
     public function storeHinhAnh(Request $request)
     {
-
+        $id_chuc_nang = 53;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         HinhAnhSanPham::create([
             'id_san_pham' => $request->id,
             'hinh_anh'    => $request->hinh_anh,
@@ -307,6 +285,15 @@ class SanPhamShopController extends Controller
     }
     public function getHinhAnh()
     {
+        $id_chuc_nang = 51;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         $check = SanPham::where('tinh_trang', 1);
         $data = SanPham::select('san_phams.id', 'san_phams.ten_san_pham')
             ->groupBy('san_phams.id', 'san_phams.ten_san_pham')
@@ -319,6 +306,15 @@ class SanPhamShopController extends Controller
 
     public function searchHinhAnh(Request $request)
     {
+        $id_chuc_nang = 55;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         $noi_dung = '%' . $request->noi_dung . '%';
 
         $data = SanPham::where('ten_san_pham', 'like', $noi_dung)
@@ -333,6 +329,15 @@ class SanPhamShopController extends Controller
 
     public function getChiTietHinhAnh(Request $request)
     {
+        $id_chuc_nang = 52;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         $id = $request->id;
         $data = HinhAnhSanPham::where('id_san_pham', $id)->get();
         if ($data->isEmpty()) {
@@ -348,6 +353,15 @@ class SanPhamShopController extends Controller
     }
     public function destroyHinhAnh(Request $request)
     {
+        $id_chuc_nang = 54;
+        $chuc_vu = Auth::guard('sanctum')->user()->id_chuc_vu;
+        $check = PhanQuyen::where('id_chuc_vu', $chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$check) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'Bạn không có quyền thực hiện chức năng này!'
+            ]);
+        }
         $id = $request->id;
         HinhAnhSanPham::where('id', $id)->delete();
